Is now able to convert OpenReq JSON received from Milla to Murmeli

+++ src/main/java/eu/openreq/mulperi/models/json/Classifier.java
@@ -40,7 +40,7 @@ public class Classifier {
 	*/
 	@SerializedName("created_at")
 	@Expose
-	private int created_at;
+	private long created_at;
 	
 	public String getId() {
 		return id;
@@ -66,11 +66,11 @@ public class Classifier {
 		this.text = text;
 	}
 	
-	public int getCreated_at() {
+	public long getCreated_at() {
 		return created_at;
 	}
 	
-	public void setCreated_at(int created_at) {
+	public void setCreated_at(long created_at) {
 		this.created_at = created_at;
 	}
 }
