+++ src/main/java/eu/openreq/mulperi/models/json/Comment.java
@@ -34,14 +34,14 @@ public class Comment {
 	*/
 	@SerializedName("created_at")
 	@Expose
-	private int created_at;
+	private long created_at;
 	/**
 	* Last modification time
 	* 
 	*/
 	@SerializedName("modified_at")
 	@Expose
-	private int modified_at;
+	private long modified_at;
 	
 	public String getId() {
 		return id;
@@ -59,19 +59,19 @@ public class Comment {
 		this.text = text;
 	}
 	
-	public int getCreated_at() {
+	public long getCreated_at() {
 		return created_at;
 	}
 	
-	public void setCreated_at(int created_at) {
+	public void setCreated_at(long created_at) {
 		this.created_at = created_at;
 	}
 	
-	public int getModified_at() {
+	public long getModified_at() {
 		return modified_at;
 	}
 	
-	public void setModified_at(int modified_at) {
+	public void setModified_at(long modified_at) {
 		this.modified_at = modified_at;
 	}
 }
