+++ src/main/java/eu/openreq/mulperi/models/json/Release.java
@@ -23,11 +23,11 @@ public class Release {
 	
 	@SerializedName("created_at")
 	@Expose
-	int created_at;
+	long created_at;
 	
 	@SerializedName("modified_at")
 	@Expose
-	int modified_at;
+	long modified_at;
 	
 	@SerializedName("start_date")
 	@Expose
@@ -42,7 +42,7 @@ public class Release {
 	 */
 	@SerializedName("capacity")
 	@Expose
-	int capacity;
+	long capacity;
 	
 	@SerializedName("project")
 	@Expose
@@ -67,19 +67,19 @@ public class Release {
 		this.id = id;
 	}
 
-	public int getCreated_at() {
+	public long getCreated_at() {
 		return created_at;
 	}
 
-	public void setCreated_at(int created_at) {
+	public void setCreated_at(long created_at) {
 		this.created_at = created_at;
 	}
 
-	public int getModified_at() {
+	public long getModified_at() {
 		return modified_at;
 	}
 
-	public void setModified_at(int modified_at) {
+	public void setModified_at(long modified_at) {
 		this.modified_at = modified_at;
 	}
 
@@ -99,11 +99,11 @@ public class Release {
 		this.release_date = release_date;
 	}
 
-	public int getCapacity() {
+	public long getCapacity() {
 		return capacity;
 	}
 
-	public void setCapacity(int capacity) {
+	public void setCapacity(long capacity) {
 		this.capacity = capacity;
 	}
